requirements.txt
# requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
python-pptx>=0.6.21
pandas>=2.0.3
python-docx>=0.8.11
numpy>=1.24.3
openai>=1.3.0
scikit-learn>=1.3.0
sentence-transformers==2.2.2